Add :node-api, factor out some artemis code
