+++ node-api/src/main/kotlin/net/corda/config/SSLConfiguration.kt
@@ -1,0 +1,12 @@
+package net.corda.config
+
+import net.corda.core.div
+import java.nio.file.Path
+
+interface SSLConfiguration {
+    val keyStorePassword: String
+    val trustStorePassword: String
+    val certificatesDirectory: Path
+    val keyStoreFile: Path get() = certificatesDirectory / "sslkeystore.jks"
+    val trustStoreFile: Path get() = certificatesDirectory / "truststore.jks"
+}
