+++ node/src/main/kotlin/net/corda/node/services/messaging/ArtemisMessagingComponent.kt
@@ -2,6 +2,7 @@ package net.corda.node.services.messaging
 
 import com.google.common.annotations.VisibleForTesting
 import com.google.common.net.HostAndPort
+import net.corda.config.SSLConfiguration
 import net.corda.core.crypto.CompositeKey
 import net.corda.core.messaging.MessageRecipientGroup
 import net.corda.core.messaging.MessageRecipients
@@ -9,14 +10,6 @@ import net.corda.core.messaging.SingleMessageRecipient
 import net.corda.core.read
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.SingletonSerializeAsToken
-import net.corda.node.services.config.SSLConfiguration
-import net.corda.node.services.messaging.ArtemisMessagingComponent.ConnectionDirection.Inbound
-import net.corda.node.services.messaging.ArtemisMessagingComponent.ConnectionDirection.Outbound
-import org.apache.activemq.artemis.api.core.TransportConfiguration
-import org.apache.activemq.artemis.core.remoting.impl.netty.NettyAcceptorFactory
-import org.apache.activemq.artemis.core.remoting.impl.netty.TransportConstants
-import java.nio.file.FileSystems
-import java.nio.file.Path
 import java.security.KeyStore
 
 /**
@@ -42,8 +35,6 @@ abstract class ArtemisMessagingComponent : SingletonSerializeAsToken() {
         const val RPC_QUEUE_REMOVALS_QUEUE = "rpc.qremovals"
         const val NOTIFICATIONS_ADDRESS = "${INTERNAL_PREFIX}activemq.notifications"
         const val NETWORK_MAP_QUEUE = "${INTERNAL_PREFIX}networkmap"
-
-        const val VERIFY_PEER_COMMON_NAME = "corda.verifyPeerCommonName"
 
         /**
          * Assuming the passed in target address is actually an ArtemisAddress will extract the host and port of the node. This should
@@ -111,19 +102,6 @@ abstract class ArtemisMessagingComponent : SingletonSerializeAsToken() {
     /** The config object is used to pass in the passwords for the certificate KeyStore and TrustStore */
     abstract val config: SSLConfiguration?
 
-    // Restrict enabled Cipher Suites to AES and GCM as minimum for the bulk cipher.
-    // Our self-generated certificates all use ECDSA for handshakes, but we allow classical RSA certificates to work
-    // in case we need to use keytool certificates in some demos
-    private val CIPHER_SUITES = listOf(
-            "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256",
-            "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256",
-            "TLS_RSA_WITH_AES_128_GCM_SHA256",
-            "TLS_ECDH_ECDSA_WITH_AES_128_GCM_SHA256",
-            "TLS_ECDH_RSA_WITH_AES_128_GCM_SHA256",
-            "TLS_DHE_RSA_WITH_AES_128_GCM_SHA256",
-            "TLS_DHE_DSS_WITH_AES_128_GCM_SHA256"
-    )
-
     /**
      * Returns nothing if the keystore was opened OK or throws if not. Useful to check the password, as
      * unfortunately Artemis tends to bury the exception when the password is wrong.
@@ -136,58 +114,5 @@ abstract class ArtemisMessagingComponent : SingletonSerializeAsToken() {
         config.trustStoreFile.read {
             KeyStore.getInstance("JKS").load(it, config.trustStorePassword.toCharArray())
         }
-    }
-
-    protected fun tcpTransport(direction: ConnectionDirection, host: String, port: Int, enableSSL: Boolean = true): TransportConfiguration {
-        // Will throw exception if enableSSL = true but config is missing
-        require(config != null || !enableSSL) { "SSL configuration cannot be null when SSL is enabled." }
-
-        val config = config
-        val options = mutableMapOf<String, Any?>(
-                // Basic TCP target details
-                TransportConstants.HOST_PROP_NAME to host,
-                TransportConstants.PORT_PROP_NAME to port,
-
-                // Turn on AMQP support, which needs the protocol jar on the classpath.
-                // Unfortunately we cannot disable core protocol as artemis only uses AMQP for interop
-                // It does not use AMQP messages for its own messages e.g. topology and heartbeats
-                // TODO further investigate how to ensure we use a well defined wire level protocol for Node to Node communications
-                TransportConstants.PROTOCOLS_PROP_NAME to "CORE,AMQP"
-        )
-
-        if (config != null && enableSSL) {
-            config.keyStoreFile.expectedOnDefaultFileSystem()
-            config.trustStoreFile.expectedOnDefaultFileSystem()
-            val tlsOptions = mapOf<String, Any?>(
-                    // Enable TLS transport layer with client certs and restrict to at least SHA256 in handshake
-                    // and AES encryption
-                    TransportConstants.SSL_ENABLED_PROP_NAME to true,
-                    TransportConstants.KEYSTORE_PROVIDER_PROP_NAME to "JKS",
-                    TransportConstants.KEYSTORE_PATH_PROP_NAME to config.keyStoreFile,
-                    TransportConstants.KEYSTORE_PASSWORD_PROP_NAME to config.keyStorePassword, // TODO proper management of keystores and password
-                    TransportConstants.TRUSTSTORE_PROVIDER_PROP_NAME to "JKS",
-                    TransportConstants.TRUSTSTORE_PATH_PROP_NAME to config.trustStoreFile,
-                    TransportConstants.TRUSTSTORE_PASSWORD_PROP_NAME to config.trustStorePassword,
-                    TransportConstants.ENABLED_CIPHER_SUITES_PROP_NAME to CIPHER_SUITES.joinToString(","),
-                    TransportConstants.ENABLED_PROTOCOLS_PROP_NAME to "TLSv1.2",
-                    TransportConstants.NEED_CLIENT_AUTH_PROP_NAME to true,
-                    VERIFY_PEER_COMMON_NAME to (direction as? Outbound)?.expectedCommonName
-            )
-            options.putAll(tlsOptions)
-        }
-        val factoryName = when (direction) {
-            is Inbound -> NettyAcceptorFactory::class.java.name
-            is Outbound -> VerifyingNettyConnectorFactory::class.java.name
-        }
-        return TransportConfiguration(factoryName, options)
-    }
-
-    protected fun Path.expectedOnDefaultFileSystem() {
-        require(fileSystem == FileSystems.getDefault()) { "Artemis only uses the default file system" }
-    }
-
-    protected sealed class ConnectionDirection {
-        object Inbound : ConnectionDirection()
-        class Outbound(val expectedCommonName: String? = null) : ConnectionDirection()
     }
 }
