+++ samples/raft-notary-demo/src/main/kotlin/net/corda/notarydemo/NotaryDemo.kt
@@ -3,6 +3,7 @@ package net.corda.notarydemo
 import com.google.common.net.HostAndPort
 import com.google.common.util.concurrent.Futures
 import joptsimple.OptionParser
+import net.corda.config.SSLConfiguration
 import net.corda.core.crypto.toStringShort
 import net.corda.core.div
 import net.corda.core.getOrThrow
@@ -10,7 +11,6 @@ import net.corda.core.messaging.CordaRPCOps
 import net.corda.core.messaging.startFlow
 import net.corda.core.transactions.SignedTransaction
 import net.corda.flows.NotaryFlow
-import net.corda.node.services.config.SSLConfiguration
 import net.corda.node.services.messaging.CordaRPCClient
 import net.corda.notarydemo.flows.DummyIssueAndMove
 import java.nio.file.Path
