+++ settings.gradle
@@ -4,6 +4,7 @@ rootProject.name = 'corda-project'
 include 'finance'
 include 'finance:isolated'
 include 'core'
+include 'node-api'
 include 'node-schemas'
 include 'node'
 include 'node:capsule'
