+++ test-utils/src/main/kotlin/net/corda/testing/messaging/SimpleMQClient.kt
@@ -1,9 +1,10 @@
 package net.corda.testing.messaging
 
 import com.google.common.net.HostAndPort
-import net.corda.node.services.config.SSLConfiguration
+import net.corda.config.SSLConfiguration
+import net.corda.node.ArtemisTcpTransport
+import net.corda.node.ConnectionDirection
 import net.corda.node.services.messaging.ArtemisMessagingComponent
-import net.corda.node.services.messaging.ArtemisMessagingComponent.ConnectionDirection.Outbound
 import net.corda.testing.configureTestSSL
 import org.apache.activemq.artemis.api.core.client.*
 
@@ -17,7 +18,7 @@ class SimpleMQClient(val target: HostAndPort,
     lateinit var producer: ClientProducer
 
     fun start(username: String? = null, password: String? = null, enableSSL: Boolean = true) {
-        val tcpTransport = tcpTransport(Outbound(), target.hostText, target.port, enableSSL)
+        val tcpTransport = ArtemisTcpTransport.tcpTransport(ConnectionDirection.Outbound(), target, config, enableSSL = enableSSL)
         val locator = ActiveMQClient.createServerLocatorWithoutHA(tcpTransport).apply {
             isBlockOnNonDurableSend = true
             threadPoolMaxSize = 1
