+++ tools/explorer/src/main/kotlin/net/corda/explorer/views/LoginView.kt
@@ -1,20 +1,13 @@
 package net.corda.explorer.views
 
 import com.google.common.net.HostAndPort
-import de.jensd.fx.glyphs.fontawesome.FontAwesomeIcon
-import de.jensd.fx.glyphs.fontawesome.FontAwesomeIconView
 import javafx.beans.property.SimpleIntegerProperty
 import javafx.scene.control.*
-import javafx.stage.FileChooser
-import net.corda.client.fxutils.map
 import net.corda.client.model.NodeMonitorModel
 import net.corda.client.model.objectProperty
 import net.corda.explorer.model.SettingsModel
-import net.corda.node.services.config.SSLConfiguration
 import org.controlsfx.dialog.ExceptionDialog
-import tornadofx.*
-import java.nio.file.Path
-import java.nio.file.Paths
+import tornadofx.View
 import kotlin.system.exitProcess
 
 class LoginView : View() {
